Adaptive conditioning, leak test and more
Add support for amplitude, period and both-modulated adaptive
conditioning to pulse helper library; creating version 2022_1.
The adaptive conditioning is most sophisticated than prior
versions because it has a lower cutoff for voltage and/or time
below which it will not descend.

Also updated the CBD helper library to version 2022_1 by adding
support for a time cutoff. I.e. the CBD process can be aborted
either by a breakdown or by a time limit. In this way, the CBD
library can be used as a crude ramp library if you set the
threshold high.

Add a leak test helper library which applys a short low-voltage
ramp and tests if the current is above threshold. If it is it
returns a bool True to represent leaky and the user can abort
their script based on it.